// src/generated/dokploy-schemas.ts
export const procedureSchemas: Record<string, ProcedureSchema> = {
// ...
  "application.saveBitbucketProvider": {
    "method": "POST",
    "path": "/application.saveBitbucketProvider",
    "tag": "application",
    "inputKind": "body",
    "inputSchema": {
      "type": "object",
      "properties": {
// ...
        "bitbucketId": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ]
        },
        "applicationId": {
          "type": "string"
        },
        "enableSubmodules": {
          "type": "boolean"
        },
        "watchPaths": {
          "anyOf": [
            {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            {
              "type": "null"
            }
          ]
        }
      },
      "required": [
        "bitbucketBranch",
        "bitbucketBuildPath",
        "bitbucketOwner",
        "bitbucketRepository",
        "bitbucketRepositorySlug",
        "bitbucketId",
        "applicationId"
      ]
    },
    "outputSchema": {
      "type": "object",
      "properties": {},
      "additionalProperties": false
    }
  },
  "application.saveGiteaProvider": {
    "method": "POST",
    "path": "/application.saveGiteaProvider",
    "tag": "application",
    "inputKind": "body",
    "inputSchema": {
      "type": "object",
      "properties": {
// ...
        "giteaId": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ]
        },
        "enableSubmodules": {
          "type": "boolean"
        },
        "watchPaths": {
          "anyOf": [
            {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            {
              "type": "null"
            }
          ]
        }
      },
      "required": [
        "applicationId",
        "giteaBranch",
        "giteaBuildPath",
        "giteaOwner",
        "giteaRepository",
        "giteaId"
      ]
    },
    "outputSchema": {
      "type": "object",
      "properties": {},
      "additionalProperties": false
    }
  },
// ...
  "application.saveGitProvider": {
    "method": "POST",
    "path": "/application.saveGitProvider",
    "tag": "application",
    "inputKind": "body",
    "inputSchema": {
      "type": "object",
      "properties": {
// ...
        "customGitUrl": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ]
        },
// ...
      },
      "required": [
        "customGitBranch",
        "applicationId",
        "customGitBuildPath",
        "customGitUrl",
        "watchPaths"
      ]
    },
    "outputSchema": {
      "type": "object",
      "properties": {},
      "additionalProperties": false
    }
  },
// ...
}